File complet

+++ src/controllers/posts.js
@@ -1,7 +1,6 @@
 const { Posts, Users } = require("../db");
-const {uploadsArchivos} = require("../cloudinary/cloudinary")
+const {uploadsArchivos, deleteArchivo} = require("../cloudinary/cloudinary")
 const fs = require("fs-extra");
-const { parse } = require("dotenv");
 const getAllPost = async (req, res) => {
   try {
     const data = await Posts.findAll({
@@ -22,43 +21,37 @@ const getAllPost = async (req, res) => {
 };
 
 const createPost = async (req, res) => {
-  console.log(req.body)
+  
   try {
     
     const { titulo, texto , userId } =await req.body;
 
     const {file} =  req.files
-    console.log(file)
+    
     if (!userId) throw new Error(" missing param id");
     const user = await Users.findByPk(userId);
-
-    const ar = await uploadsArchivos(file.tempFilePath)
-      ///////////////////////////////// req.body
-    
-     console.log("*******************************")
-     console.log(ar)
-     console.log("*******************************")
-      let cosita =  ar.url
-      let fotos= ar.secure_url
-
     if (!user) throw new Error("No se encuentra el usuario");
-    const newPost = await Posts.create({
-      titulo,
-      texto,
-      userId,
-      media: cosita,
-      foto: fotos
-    });
-
-    if (!newPost) throw new Error("No se pudo crear el post");
-
-    user.addPosts(newPost);
-    
-     await fs.unlink(req.files.file.tempFilePath)
-    res.status(200).send({
-      msg: "Post Creado Exitosamente",
-      post: newPost,
-    });
+    if(file){
+      const ar = await uploadsArchivos(file.tempFilePath)
+        let cosita =  ar.url
+        let fotos= ar.secure_url
+      const newPost = await Posts.create({
+        titulo,
+        texto,
+        userId,
+        media: cosita,
+        foto: fotos
+      });
+      if (!newPost) throw new Error("No se pudo crear el post");
+  
+      user.addPosts(newPost);
+      
+      await fs.unlink(req.files.file.tempFilePath)
+      res.status(200).send({
+        msg: "Post Creado Exitosamente",
+        post: newPost,
+      });
+    }
   } catch (err) {
     res.status(500).send({ msg: "Erorr en el servidor: ", err: err.message });
   }
@@ -83,12 +76,17 @@ const eliminarPost = async (req, res) => {
   try {
     let { id } = req.params;
     let buscarid = await Posts.findByPk(id);
+    if (buscarid.media.length > 15){
+      await deleteArchivo(buscarid.media)
+    }
     if (buscarid) {
+
       await Posts.destroy({
         where: {
           id: id,
         },
       });
+      
       res.status(200).json({ msg: "Se elimino el posteo" });
     } else {
       res.status(400).json({ msg: "pasame un id capo" });
